refactor: renamed internal/environment package to internal/featureflags
